+++ inventory-management/src/pages/production/ProductionActionBar.tsx
@@ -1,0 +1,235 @@
+import React from "react";
+import {
+  Search,
+  Filter,
+  ChevronDown,
+  ChevronUp,
+  Plus,
+  RefreshCw,
+  Download,
+  FileText,
+  XCircle,
+} from "lucide-react";
+import { Product, Warehouse } from "./types";
+
+export interface FilterParams {
+  productId?: string | number;
+  mainProductId?: string | number;
+  startDate?: string;
+  endDate?: string;
+  warehouseId?: string | number;
+  status?: string;
+}
+
+interface ProductionActionBarProps {
+  searchTerm: string;
+  onSearch: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  onSearchSubmit: (e: React.FormEvent) => void;
+  showFilters: boolean;
+  onToggleFilters: () => void;
+  viewType: "table" | "cards";
+  onToggleViewType: () => void;
+  onRefresh: () => void;
+  onAddClick: () => void;
+  products: Product[];
+  warehouses: Warehouse[];
+  filters: FilterParams;
+  onFilterChange: (
+    e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>
+  ) => void;
+  onApplyFilters: () => void;
+  onClearFilters: () => void;
+}
+
+const ProductionActionBar: React.FC<ProductionActionBarProps> = ({
+  searchTerm,
+  onSearch,
+  onSearchSubmit,
+  showFilters,
+  onToggleFilters,
+  viewType,
+  onToggleViewType,
+  onRefresh,
+  onAddClick,
+  products,
+  filters,
+  onFilterChange,
+  onApplyFilters,
+  onClearFilters,
+}) => {
+  return (
+    <div className="bg-white rounded-xl shadow-sm border border-gray-200 p-4 md:p-6 mb-6 md:mb-8">
+      <div className="flex flex-col sm:flex-row justify-between items-start sm:items-center gap-4">
+        {/* Search Input */}
+        <form
+          onSubmit={onSearchSubmit}
+          className="relative flex-1 w-full max-w-md"
+        >
+          <Search
+            className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400"
+            size={18}
+          />
+          <input
+            type="text"
+            placeholder="Search by reference, product, notes..."
+            className="w-full pl-10 pr-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500"
+            value={searchTerm}
+            onChange={onSearch}
+          />
+        </form>
+
+        {/* Action Buttons */}
+        <div className="flex flex-wrap gap-2 w-full sm:w-auto">
+          <button
+            onClick={onToggleFilters}
+            className="flex items-center px-3 md:px-4 py-2 border border-gray-300 rounded-lg hover:bg-gray-50 text-sm"
+            aria-expanded={showFilters}
+            aria-controls="filters-panel"
+          >
+            <Filter size={16} className="mr-1 md:mr-2" />
+            <span>Filters</span>
+            {showFilters ? (
+              <ChevronUp size={16} className="ml-1" />
+            ) : (
+              <ChevronDown size={16} className="ml-1" />
+            )}
+          </button>
+
+          <button
+            onClick={onToggleViewType}
+            className="flex items-center px-3 md:px-4 py-2 border border-gray-300 rounded-lg hover:bg-gray-50 text-sm"
+            title={`Switch to ${viewType === "table" ? "card" : "table"} view`}
+          >
+            <FileText size={16} className="mr-1" />
+            <span>{viewType === "table" ? "Cards" : "Table"}</span>
+          </button>
+
+          <button
+            onClick={() => console.log("Export feature coming soon!")}
+            className="flex items-center px-3 md:px-4 py-2 border border-gray-300 rounded-lg hover:bg-gray-50 text-sm"
+            title="Export data"
+          >
+            <Download size={16} className="mr-1" />
+            <span>Export</span>
+          </button>
+
+          <button
+            onClick={onRefresh}
+            className="flex items-center px-3 md:px-4 py-2 border border-gray-300 rounded-lg hover:bg-gray-50 text-sm"
+            title="Refresh data"
+          >
+            <RefreshCw size={16} />
+            <span className="sr-only">Refresh</span>
+          </button>
+
+          <button
+            onClick={onAddClick}
+            className="flex items-center px-3 md:px-4 py-2 bg-blue-600 text-white rounded-lg hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-blue-500 text-sm font-medium"
+          >
+            <Plus size={16} className="mr-1 md:mr-2" />
+            <span>New Batch</span>
+          </button>
+        </div>
+      </div>
+
+      {/* Filters Panel */}
+      {showFilters && (
+        <div
+          id="filters-panel"
+          className="mt-4 p-4 bg-gray-50 rounded-lg border border-gray-200 transition-all"
+        >
+          <div className="flex justify-between items-center mb-3">
+            <h3 className="text-sm font-medium text-gray-700 flex items-center">
+              <Filter size={16} className="mr-2" />
+              Filters
+            </h3>
+            <button
+              onClick={onClearFilters}
+              className="text-sm text-blue-600 hover:text-blue-800 flex items-center gap-1"
+            >
+              <XCircle size={16} />
+              Clear Filters
+            </button>
+          </div>
+
+          <div className="grid grid-cols-1 sm:grid-cols-3 gap-4">
+            {/* Product Filter */}
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-1">
+                Produced Product
+              </label>
+              <select
+                name="productId"
+                value={filters.productId?.toString() || ""}
+                className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 text-sm"
+              >
+                <option value="">All Products</option>
+                {products.map((product) => (
+                  <option key={product.id} value={product.id}>
+                    {product.name}
+                  </option>
+                ))}
+              </select>
+            </div>
+
+            {/* Raw Material Filter */}
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-1">
+                Raw Material
+              </label>
+              <select
+                name="mainProductId"
+                value={filters.mainProductId || ""}
+                onChange={onFilterChange}
+                className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 text-sm"
+              >
+                <option value="">All Raw Materials</option>
+                {products.map((product) => (
+                  <option key={product.id} value={product.id}>
+                    {product.name}
+                  </option>
+                ))}
+              </select>
+            </div>
+
+            {/* Date Range Filter */}
+            <div>
+              <label className="block text-sm font-medium text-gray-700 mb-1">
+                Date Range
+              </label>
+              <div className="grid grid-cols-2 gap-2">
+                <input
+                  type="date"
+                  name="startDate"
+                  value={filters.startDate || ""}
+                  onChange={onFilterChange}
+                  className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 text-sm"
+                  placeholder="Start Date"
+                />
+                <input
+                  type="date"
+                  name="endDate"
+                  value={filters.endDate || ""}
+                  onChange={onFilterChange}
+                  className="w-full px-3 py-2 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500 text-sm"
+                  placeholder="End Date"
+                />
+              </div>
+            </div>
+          </div>
+
+          <div className="mt-4 flex justify-end">
+            <button
+              onClick={onApplyFilters}
+              className="px-4 py-2 bg-blue-600 text-white rounded-md text-sm font-medium hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-blue-500"
+            >
+              Apply Filters
+            </button>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+};
+
+export default ProductionActionBar;
