+++ inventory-management/src/pages/production/ProductionTable.tsx
@@ -1,0 +1,306 @@
+import React from "react";
+import { Eye, Edit2, Trash2, ArrowLeft, ArrowRight } from "lucide-react";
+import { Production } from "./types";
+import { formatCurrency, formatDate, formatNumber } from "./utils";
+
+interface ProductionTableProps {
+  productions: Production[];
+  loading: boolean;
+  error: string | null;
+  page: number;
+  totalPages: number;
+  totalProductions: number;
+  pageSize: number;
+  sortConfig: { key: string; direction: string } | null;
+  onPageChange: (newPage: number) => void;
+  onSort: (key: string) => void;
+  onView: (production: Production) => void;
+  onEdit: (production: Production) => void;
+  onDelete: (productionId: number) => void;
+}
+
+const ProductionTable: React.FC<ProductionTableProps> = ({
+  productions,
+  loading,
+  error,
+  page,
+  totalPages,
+  totalProductions,
+  pageSize,
+  sortConfig,
+  onPageChange,
+  onSort,
+  onView,
+  onEdit,
+  onDelete,
+}) => {
+  const columns = [
+    { key: "referenceNumber", label: "Reference", sortable: true },
+    { key: "product", label: "Product", sortable: true },
+    { key: "quantityProduced", label: "Quantity", sortable: true },
+    { key: "date", label: "Date", sortable: true },
+    { key: "unitPrice", label: "Unit Price", sortable: true },
+    { key: "actions", label: "Actions", sortable: false },
+  ];
+
+// 
+
+  return (
+    <div className="bg-white rounded-xl shadow-sm border border-gray-200 overflow-hidden">
+      <div className="overflow-x-auto">
+        <table className="min-w-full divide-y divide-gray-200">
+          <thead className="bg-gray-50">
+            <tr>
+              {columns.map((column) => (
+                <th
+                  key={column.key}
+                  className={`px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider ${
+                    column.sortable ? "cursor-pointer hover:bg-gray-100" : ""
+                  }`}
+                  onClick={() => column.sortable && onSort(column.key)}
+                >
+                  <div className="flex items-center">
+                    {column.label}
+                    {sortConfig?.key === column.key && (
+                      <span className="ml-1">
+                        {sortConfig.direction === "ascending" ? "↑" : "↓"}
+                      </span>
+                    )}
+                  </div>
+                </th>
+              ))}
+            </tr>
+          </thead>
+          <tbody className="bg-white divide-y divide-gray-200">
+            {loading ? (
+              // Skeleton rows
+              Array(5).fill(0).map((_, i) => (
+                <tr key={`skeleton-${i}`} className="animate-pulse">
+                  <td className="px-6 py-4">
+                    <div className="h-4 bg-gray-200 rounded w-3/4 mb-2"></div>
+                    <div className="h-3 bg-gray-200 rounded w-1/2"></div>
+                  </td>
+                  <td className="px-6 py-4">
+                    <div className="h-4 bg-gray-200 rounded w-2/3"></div>
+                  </td>
+                  <td className="px-6 py-4">
+                    <div className="h-4 bg-gray-200 rounded w-3/4"></div>
+                  </td>
+                  <td className="px-6 py-4">
+                    <div className="h-4 bg-gray-200 rounded w-1/2"></div>
+                  </td>
+                  <td className="px-6 py-4">
+                    <div className="h-4 bg-gray-200 rounded w-2/3"></div>
+                  </td>
+                  <td className="px-6 py-4 text-right">
+                    <div className="h-6 bg-gray-200 rounded w-16 inline-block"></div>
+                  </td>
+                </tr>
+              ))
+            ) : error ? (
+              <tr>
+                <td colSpan={columns.length} className="px-6 py-4 text-center">
+                  <div className="flex items-center justify-center text-red-600">
+                    {error}
+                  </div>
+                </td>
+              </tr>
+            ) : productions.length === 0 ? (
+              <tr>
+                <td colSpan={columns.length} className="px-6 py-4 text-center text-gray-500">
+                  No production batches found.
+                </td>
+              </tr>
+            ) : (
+              productions.map((production) => (
+                <tr key={production.id} className="hover:bg-gray-50 transition-colors">
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="text-sm font-medium text-blue-600">
+                      {production.referenceNumber}
+                    </div>
+                    <div className="text-xs text-gray-500">
+                      {formatDate(production.date)}
+                    </div>
+                  </td>
+
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="text-sm font-medium">
+                      {production.product?.name || "N/A"}
+                    </div>
+                    {production.mainProduct && (
+                      <div className="text-xs text-gray-500">
+                        From: {production.mainProduct.name}
+                      </div>
+                    )}
+                  </td>
+
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="text-sm font-medium">
+                      {formatNumber(production.quantityProduced)} Kg
+                    </div>
+                    {production.usedQuantity && (
+                      <div className="text-xs text-gray-500">
+                        Used: {formatNumber(production.usedQuantity)} Kg
+                      </div>
+                    )}
+                  </td>
+
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="text-sm">
+                      {formatDate(production.date)}
+                    </div>
+                  </td>
+
+                  <td className="px-6 py-4 whitespace-nowrap">
+                    <div className="text-sm font-medium">
+                      {production.unitPrice ? formatCurrency(production.unitPrice) : "N/A"}
+                    </div>
+                  </td>
+
+                  <td className="px-6 py-4 whitespace-nowrap text-right text-sm font-medium">
+                    <div className="flex items-center justify-end space-x-2">
+                      <button
+                        onClick={() => onView(production)}
+                        className="text-green-600 hover:text-green-900 p-1 rounded-full hover:bg-green-50"
+                        title="View Details"
+                      >
+                        <Eye size={18} />
+                      </button>
+
+                      <button
+                        onClick={() => onEdit(production)}
+                        className="text-blue-600 hover:text-blue-900 p-1 rounded-full hover:bg-blue-50"
+                        title="Edit"
+                      >
+                        <Edit2 size={18} />
+                      </button>
+
+                      <button
+                        onClick={() => onDelete(production.id)}
+                        className="text-red-600 hover:text-red-900 p-1 rounded-full hover:bg-red-50"
+                        title="Delete"
+                      >
+                        <Trash2 size={18} />
+                      </button>
+                    </div>
+                  </td>
+                </tr>
+              ))
+            )}
+          </tbody>
+        </table>
+      </div>
+
+      {productions.length > 0 && (
+        <div className="bg-gray-50 px-4 py-3 flex items-center justify-between border-t border-gray-200 sm:px-6">
+          <div className="flex-1 flex justify-between sm:hidden">
+            <button
+              onClick={() => onPageChange(page - 1)}
+              disabled={page === 1}
+              className={`relative inline-flex items-center px-4 py-2 border border-gray-300 text-sm font-medium rounded-md ${
+                page === 1
+                  ? "bg-gray-100 text-gray-400 cursor-not-allowed"
+                  : "bg-white text-gray-700 hover:bg-gray-50"
+              }`}
+            >
+              <ArrowLeft size={16} className="mr-1" />
+              Previous
+            </button>
+            <span className="text-sm text-gray-700">
+              Page {page} of {totalPages}
+            </span>
+            <button
+              onClick={() => onPageChange(page + 1)}
+              disabled={page >= totalPages}
+              className={`relative inline-flex items-center px-4 py-2 border border-gray-300 text-sm font-medium rounded-md ${
+                page >= totalPages
+                  ? "bg-gray-100 text-gray-400 cursor-not-allowed"
+                  : "bg-white text-gray-700 hover:bg-gray-50"
+              }`}
+            >
+              Next
+              <ArrowRight size={16} className="ml-1" />
+            </button>
+          </div>
+          <div className="hidden sm:flex-1 sm:flex sm:items-center sm:justify-between">
+            <div>
+              <p className="text-sm text-gray-700">
+                Showing{" "}
+                <span className="font-medium">
+                  {(page - 1) * pageSize + 1}
+                </span>{" "}
+                to{" "}
+                <span className="font-medium">
+                  {Math.min(page * pageSize, totalProductions)}
+                </span>{" "}
+                of <span className="font-medium">{totalProductions}</span>{" "}
+                results
+              </p>
+            </div>
+            <div>
+              <nav
+                className="relative z-0 inline-flex rounded-md shadow-sm -space-x-px"
+                aria-label="Pagination"
+              >
+                <button
+                  onClick={() => onPageChange(page - 1)}
+                  disabled={page === 1}
+                  className={`relative inline-flex items-center px-2 py-2 rounded-l-md border border-gray-300 bg-white text-sm font-medium ${
+                    page === 1
+                      ? "text-gray-300 cursor-not-allowed"
+                      : "text-gray-500 hover:bg-gray-50"
+                  }`}
+                >
+                  <span className="sr-only">Previous</span>
+                  <ArrowLeft className="h-5 w-5" aria-hidden="true" />
+                </button>
+
+                {/* Page Numbers */}
+                {Array.from({ length: Math.min(5, totalPages) }, (_, i) => {
+                  let pageNum;
+                  if (totalPages <= 5) {
+                    pageNum = i + 1;
+                  } else if (page <= 3) {
+                    pageNum = i + 1;
+                  } else if (page >= totalPages - 2) {
+                    pageNum = totalPages - 4 + i;
+                  } else {
+                    pageNum = page - 2 + i;
+                  }
+                  return (
+                    <button
+                      key={pageNum}
+                      onClick={() => onPageChange(pageNum)}
+                      className={`relative inline-flex items-center px-4 py-2 border text-sm font-medium ${
+                        page === pageNum
+                          ? "z-10 bg-blue-50 border-blue-500 text-blue-600"
+                          : "bg-white border-gray-300 text-gray-500 hover:bg-gray-50"
+                      }`}
+                    >
+                      {pageNum}
+                    </button>
+                  );
+                })}
+
+                <button
+                  onClick={() => onPageChange(page + 1)}
+                  disabled={page >= totalPages}
+                  className={`relative inline-flex items-center px-2 py-2 rounded-r-md border border-gray-300 bg-white text-sm font-medium ${
+                    page >= totalPages
+                      ? "text-gray-300 cursor-not-allowed"
+                      : "text-gray-500 hover:bg-gray-50"
+                  }`}
+                >
+                  <span className="sr-only">Next</span>
+                  <ArrowRight className="h-5 w-5" aria-hidden="true" />
+                </button>
+              </nav>
+            </div>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+};
+
+export default ProductionTable;
